+++ sputofy_2/lib/components/playlistsPage/show_hidden_playlists.dart
@@ -1,7 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
-import 'package:sputofy_2/models/playlist_model.dart';
+
 import 'package:sputofy_2/providers/provider.dart';
+import 'package:provider/provider.dart';
+
+import 'package:sputofy_2/models/playlist_model.dart';
 
 void showHiddenPlaylist(BuildContext context,
     {List<Playlist>? importedPlaylists}) {
