+++ sputofy_2/pubspec.yaml
@@ -30,13 +30,16 @@ dependencies:
   cupertino_icons: ^1.0.2
 
   file_picker: ^3.0.3
+  image_cropper: ^1.4.1
+  
   provider: ^5.0.0
-  sqflite: ^2.0.0+3
   path_provider: ^2.0.2
+  
+  sqflite: ^2.0.0+3
+  
   permission_handler: ^8.1.1
   shared_preferences: ^2.0.6
 
-  flutter_svg: ^0.22.0
 
   just_audio: ^0.8.0
   audio_service: ^0.17.1
@@ -45,8 +48,6 @@ dependencies:
   youtube_explode_dart: ^1.9.9
   share_plus: ^2.1.4
 
-  image_picker: ^0.8.2
-  image_cropper: ^1.4.1
 
 
 dev_dependencies:
